refactor import for mapreduce.Peer
