+++ mapreduce/__init__.py
@@ -24,4 +24,4 @@ __spec_version__ = (1000 * int(version_split[0])) + (10 * int(version_split[1]))
 # Import all submodules.
 from . import protocol
 from . import utils
-from . import peer
+from .peer import Peer
